Current version is a dice roller

+++ playground/main.cpp
@@ -4,8 +4,8 @@
 using namespace std;
 int main() {
     srand(time(0));
-    const int maxValue = 6;
-    const int minValue = 1;
-    int result = (rand() % (maxValue - minValue + 1)) + minValue;
+    const short maxValue = 6;
+    const short minValue = 1;
+    int result = (rand() % (maxValue - minValue) + 1) + minValue;
     cout << "Result: " << result << endl;
 }
